Added new program Propagate_Orbits, updated Orbitography_Functions
+ Propagate_Orbits.py loads data from the satellite database and
and propagates the orbits forward in time. Calculates the position
of the satellites at predetermined points in time, along with some
additional data for plotting purposes

+ Updated Orbitography_Functions.py to accept new input data format
from Propagate_Orbits.py. Not all functions are yet updated.

+ Updated .gitignore

TODO:
- Modify Propagate_Orbits.py to accept additional information required
for numerical propagation

- Update outdated functions in Orbitography_Functions.py

- Modify SatelliteFunctions.propagateTLE (and propagateNumerical) to
output the list of datetimes that the orbits are calculated at. Save
this data in a seperate plotSettings.pkl file and remove from
prop_data database, to improve speed and reduce bloat